packages/op-coreutils/pyproject.toml
[project]
name = "op-coreutils"
version = "0.1.0"
description = "Core utitilies used across all op-analytics packages."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.43",
    "google-cloud-storage>=2.18.2",
    "structlog>=24.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
